Test
This project is developed in C# (Visual Studio 2013).I will accept
anyone with basic c# knowledge, and (optional) design (visual studio).
This is a simple project to get yourself off the computer. This project
will be for you if you want to practice multi-threading, stopwatches,
working with files, try, if, while loops and more! You will not get in
any way paid in this project, it is just a fun project to practice
skills. This project is not going to have ads (unless the project forces
it to.).